requirements.txt
dj-database-url==0.4.1
Django==1.11.1
djangorestframework==3.7.3
gunicorn==19.6.0
psycopg2==2.7.3.1
pyactiveresource==2.1.2
pytz==2017.3
PyYAML==3.12
ShopifyAPI==2.5.1
six==1.11.0
whitenoise==3.2
